pyproject.toml
[project]
name = "drf-tutorial"
version = "0.0.1"
requires-python = ">=3.14"

dependencies = [
    "django~=5.0",
    "djangorestframework",
    "pygments"
]

[dependency-groups]
lint = [
    "ruff",
    "mypy",
    "django-stubs[compatible-mypy]",
    "django-stubs-ext",
    "djangorestframework-stubs[compatible-mypy]",
    "types-Pygments"
]

[tool.uv]
default-groups = ["lint"]

[tool.mypy]
# https://mypy.readthedocs.io/en/stable/command_line.html#cmdoption-mypy-exclude
exclude = [
    ".venv",
    "build",
    "manage.py"
]
ignore_errors = false
warn_return_any = true
disallow_untyped_defs = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[[tool.mypy.overrides]]
module = ["*.migrations.*"]
ignore_errors = true

[tool.django-stubs]
# Intentionally left empty, `django_settings_module` will be read from DJANGO_SETTINGS_MODULE env var.
# https://github.com/typeddjango/django-stubs/issues/2912

[tool.ruff]
line-length = 100
extend-exclude = [
    "build",
    "*.egg-info",
    "migrations"
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
